Move test to DataFormatConvertersTest
